+++ src/test/java/StepDefinitions/Home/US_3_HomeSteps.java
@@ -1,8 +1,8 @@
 package StepDefinitions.Home;
 
 import Pages.DialogContent_US_3;
+
 import Pages.LeftNav_US_3;
-import StepDefinitions.Hooks;
 import Utilities.GWD;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Given;
@@ -15,7 +15,7 @@ import java.util.List;
 
 public class US_3_HomeSteps {
 
-    DialogContent_US_3 dc=new DialogContent_US_3();
+    LeftNav_US_3 ln=new LeftNav_US_3();
 
 
     @Given("Navigate to Campus")
@@ -25,9 +25,9 @@ public class US_3_HomeSteps {
 
     @When("Enter username {string} and password {string} and click login button")
     public void enterUsernameAndPasswordAndClickLoginButton(String username, String password) {
-        dc.mySendKeys(dc.username, username);
-        dc.mySendKeys(dc.password, password);
-        dc.myClick(dc.loginButton);
+        ln.mySendKeys(ln.username, username);
+        ln.mySendKeys(ln.password, password);
+        ln.myClick(ln.loginButton);
     }
 
     @Given("Click on all buttons in the tab menu")
@@ -35,8 +35,8 @@ public class US_3_HomeSteps {
         List<String> tabMenuButtonList=tabMenuButtons.asList(String.class);
 
         for (int i = 0; i < tabMenuButtonList.size(); i++) {
-            WebElement webElement = dc.getWebelement(tabMenuButtonList.get(i));
-            dc.myClick(webElement);
+            WebElement webElement = ln.getWebelement(tabMenuButtonList.get(i));
+            ln.myClick(webElement);
             new Actions(GWD.getDriver()).sendKeys(Keys.ESCAPE).build().perform();
             Thread.sleep(3000);
